# Conflicts:
#	src/main/java/pl/coderslab/controller/HomeController.java
#	src/main/java/pl/coderslab/repository/RecipeRepository.java
#	src/main/java/pl/coderslab/service/RecipeService.java

+++ src/main/java/pl/coderslab/repository/PlanRepository.java
@@ -13,8 +13,8 @@ public interface PlanRepository extends JpaRepository<Plan, Long>
     int getUserPlansCount(Long userId);
 
     @Query(nativeQuery = true, name = "LastPlanResult")
-    List<RecipePlanDTO> usrLstPln(int userId); // mapped to RecipePlanDTO in Plan entity
+    List<RecipePlanDTO> usrLstPln(Long userId); // mapped to RecipePlanDTO in Plan entity
 
     @Query(value = "SELECT name FROM plan WHERE id = (SELECT MAX(id) FROM plan WHERE admin_id = ?1)", nativeQuery = true)
-    String getLastPlanName(int userId);
+    String getLastPlanName(Long userId);
 }
